feat: InfluxDB 센서 평균값 조회 API 추가 및 소수점 정밀도 개선
feat: InfluxDB 센서 평균값 조회 API 추가 및 소수점 정밀도 개선

- 24시간 동안의 센서 데이터 평균값을 조회하는 API 엔드포인트를 추가합니다.
- InfluxDB에서 조회된 센서 값의 데이터 타입을 Double에서 BigDecimal로 변경하여 소수점 정밀도를 높였습니다.
- 센서 트렌드 조회 시 응답에 cultivationId, deviceEui, sensorType을 포함하도록 수정했습니다.
- 관련 테스트 코드를 수정된 로직에 맞게 업데이트했습니다.

+++ src/main/java/site/yesaido/cultivation_server/sensor/dto/response/influx/SensorTypeAverageResponse.java
@@ -1,9 +1,11 @@
 package site.yesaido.cultivation_server.sensor.dto.response.influx;
 
+import java.math.BigDecimal;
+
 public record SensorTypeAverageResponse(
         Long cultivationId,
         String sensorType,
         String unit,
-        Double averageValue
+        BigDecimal averageValue
 ) {
 }
